Extracts data that's displayed in other parts of the site into their own components to show in the ipad view
